% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep_eda_ira_per.R
\name{basal_ira_sp}
\alias{basal_ira_sp}
\alias{basal_eda_sp}
\title{Limpiar bases de EDA e IRA}
\usage{
basal_ira_sp(rute_ira, year_range = 2013:2019, diagnosis = "ira")

basal_eda_sp(rute_eda, year_range = 2013:2019, diagnosis = "eda")
}
\arguments{
\item{rute_ira}{ruta a la carpeta con todas las bases de datos crudas de IRA}

\item{year_range}{rango de años a usar como basal}

\item{diagnosis}{elegir el diagnostico a evaluar. para IRA: ira, neu, ngr, hos, dih, deh, sob. para EDA: eda, col, dis, cop.}

\item{rute_eda}{ruta a la carpeta con todas las bases de datos crudas de EDA}
}
\value{
Bases limpiar de EDA e IRA.
}
\description{
Funciones para limpiar bases de EDA e IRA.
}
\section{Functions}{
\itemize{
\item \code{basal_ira_sp}: limpiar base IRA

\item \code{basal_eda_sp}: remove replicates
}}

\examples{

# to run this examples
# change rute_eda or rute_ira accordingly
#
# library(sitreper)
# library(tidyverse)
#
# #ira
# diagnosis <- "neu" #ira, neu, ngr, hos, dih, deh, sob
# ira_ira <- basal_ira_sp(rute_ira = "data_raw/ira",
#                         year_range = 2018:2019,
#                         diagnosis = diagnosis)
# ira_ira \%>\% count(ano)
# ira_ira \%>\% count(semana)
#
# #eda
# diagnosis <- "daa" #col, dis, cop
# eda_eda <- basal_eda_sp(rute_eda = "data_raw/eda",
#                         year_range = 2018:2019,
#                         diagnosis = diagnosis)
# eda_eda \%>\% count(type,ano)
# eda_eda \%>\% count(semana)

}
