% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgb_surveillance.R
\name{sgb_replicate_filter}
\alias{sgb_replicate_filter}
\alias{sgb_replicate_remove}
\title{sgb surveillance specific functions}
\usage{
sgb_replicate_filter(data, output_rute)

sgb_replicate_remove(data, input_rute)
}
\arguments{
\item{data}{raw data input file}

\item{output_rute}{output rute name}

\item{input_rute}{input of reviewed rpelicate output with column "eliminar_casos"}
}
\value{
replicates identification and removal functions
}
\description{
clean sgb surveillance dataset
}
\section{Functions}{
\itemize{
\item \code{sgb_replicate_filter}: filter replicates and store them in excel file

\item \code{sgb_replicate_remove}: remove replicates
}}

\examples{


}
