% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep_delimit.R
\name{sitrep_delimit_metadata}
\alias{sitrep_delimit_metadata}
\alias{sitrep_delimit_spacetime_eval}
\alias{sitrep_delimit_spacetime_make}
\alias{sitrep_delimit_spacetime_eval_free}
\title{Delimit in space and time}
\usage{
sitrep_delimit_metadata(data_raw_name_string, week_ini = "2020-W01",
  week_fin = "today", space_num = 4)

sitrep_delimit_spacetime_eval(data, week_ini = "2020-W01",
  week_fin = "today", space_var, space_num = 4)

sitrep_delimit_spacetime_make(data, week_ini = "2020-W01",
  week_fin = "today", space_var, space_num = 4)

sitrep_delimit_spacetime_eval_free(data, ..., week_ini = "2020-W01",
  week_fin = "today", space_num = 4)
}
\arguments{
\item{data_raw_name_string}{filename as string "data-raw/ficha_GB_26Feb20_09_52.xls"}

\item{week_ini}{week ini}

\item{week_fin}{week fin}

\item{space_num}{number of administrative units}

\item{data}{outcome}

\item{space_var}{define the administrative unit variable}

\item{...}{more spatial covariates}

\item{data}{outcome}

\item{data}{outcome}
}
\value{
retorno
}
\description{
Create
}
\section{Functions}{
\itemize{
\item \code{sitrep_delimit_metadata}: 

\item \code{sitrep_delimit_spacetime_eval}: create 2

\item \code{sitrep_delimit_spacetime_make}: create 3

\item \code{sitrep_delimit_spacetime_eval_free}: create 2 free
}}

\examples{


}
