% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep_geocode.R
\name{sitrep_opencage_tidy}
\alias{sitrep_opencage_tidy}
\alias{sitrep_opencage_tidy_map}
\alias{sitrep_opencage_create}
\title{apply a tidy geocoding with fixed parameters}
\usage{
sitrep_opencage_tidy(place, opencagekey)

sitrep_opencage_tidy_map(data, place, opencagekey)

sitrep_opencage_create(data, address, administrative_level)
}
\arguments{
\item{place}{string with place name}

\item{opencagekey}{opencage key}

\item{data}{tibble with column of places}

\item{address}{address to geocode}

\item{administrative_level}{adminsitrative level to add to the string}

\item{data}{tibble with column of places and an administrative level}
}
\value{
columns with geolocation coordinates
}
\description{
geocode a vector inside a tibble using opencage
}
\section{Functions}{
\itemize{
\item \code{sitrep_opencage_tidy}: tidy opencage output for only one string

\item \code{sitrep_opencage_tidy_map}: tidy opencage output for only a tibble

\item \code{sitrep_opencage_create}: creates clean column previous to start geocoding
}}

\examples{

#
#library(opencage)
#library(tidyverse)
#library(sitreper)
#
#opencagekey <- "paste_free_geocode_key_here"
#
## examples ----------------------------------------------------------------
#
#place_01 <- "Calle Daniel Olaechea 199, LIMA PERU"
#place_02 <- "Ministerio de Salud Lima Peru"
#place_03 <- "Av. Salaverry 801, Jesus Maria 15072"
#place_10 <- "Sarzeau"
#
## -------------------------------------------------------------------------
#
#output_clean <- sitrep_opencage_tidy(place = place_02,opencagekey = opencagekey)
#
#output_clean \%>\% glimpse()
##> Observations: 1
##> Variables: 22
##> $ `components.ISO_3166-1_alpha-2` <fct> PE
##> $ `components.ISO_3166-1_alpha-3` <fct> PER
##> $ components._category            <fct> building
##> $ components._type                <fct> building
##> $ components.city                 <fct> Jesús María
##> $ components.continent            <fct> South America
##> $ components.country              <fct> Perú
##> $ components.country_code         <fct> pe
##> $ components.house_number         <fct> 801
##> $ components.postcode             <fct> LIMA 15046
##> $ components.region               <fct> Lima
##> $ components.road                 <fct> Avenida Salaverry
##> $ components.state                <fct> Lima
##> $ components.state_code           <fct> LIM
##> $ components.suburb               <fct> Jesús María
##> $ components.unknown              <fct> Ministerio de Salud
##> $ geometry.lat                    <dbl> -12.07298
##> $ geometry.lng                    <dbl> -77.04095
##> $ confidence                      <fct> 10
##> $ formatted                       <fct> "Ministerio de Salud, Avenida ...
##> $ query                           <chr> "Ministerio de Salud Lima Peru"
##> $ annotations.OSM.url             <fct> https://www.openstreetmap.org/...
#
## -------------------------------------------------------------------------
#
#data_query <- tibble(place=c(place_01,place_02,place_03,place_10))
#
#data_query \%>\%
#  sitrep_opencage_tidy_map(place = place,opencagekey = opencagekey)
##> # A tibble: 3 x 8
##>   place geocode geometry.lat geometry.lng confidence components.post~
##>   <chr> <list>         <dbl>        <dbl> <fct>      <fct>
##> 1 Call~ <tibbl~        -12.1        -77.0 9          15072
##> 2 Mini~ <tibbl~        -12.1        -77.0 10         LIMA 15046
##> 3 Av. ~ <tibbl~        -12.1        -77.0 10         LIMA 15046
##> # ... with 2 more variables: components._category <fct>,
##> #   annotations.OSM.url <fct>

}
