% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep_delimit.R
\name{sitrep_delimit_metadata}
\alias{sitrep_delimit_metadata}
\alias{sitrep_delimit_spacetime_eval}
\alias{sitrep_delimit_spacetime_make}
\alias{sitrep_delimit_spacetime_eval_free}
\alias{sitrep_delimit_spacetime_count}
\alias{sitrep_delimit_spacetime_filter}
\alias{sitrep_delimit_spacetime_count_free}
\title{Delimit in space and time}
\usage{
sitrep_delimit_metadata(
  data_raw_name_string,
  week_ini = "2020-W01",
  week_fin = "today",
  space_num = 4
)

sitrep_delimit_spacetime_eval(
  data,
  week_ini = "2020-W01",
  week_fin = "today",
  space_var,
  space_num = 4
)

sitrep_delimit_spacetime_make(
  data,
  week_ini = "2020-W01",
  week_fin = "today",
  space_var,
  space_num = 4
)

sitrep_delimit_spacetime_eval_free(
  data,
  ...,
  week_ini = "2020-W01",
  week_fin = "today",
  space_num = 4
)

sitrep_delimit_spacetime_count(
  data,
  week_ini = "2020-W01",
  week_fin = "today",
  space_var,
  space_num = 4
)

sitrep_delimit_spacetime_filter(
  data,
  week_var,
  week_ini = "2020-W01",
  week_fin = "today",
  space_var,
  space_num = 4
)

sitrep_delimit_spacetime_count_free(
  data,
  ...,
  week_var,
  week_ini = "2020-W01",
  week_fin = "today",
  space_num = 4
)
}
\arguments{
\item{data_raw_name_string}{filename as string "data-raw/ficha_GB_26Feb20_09_52.xls"}

\item{week_ini}{week ini}

\item{week_fin}{week fin}

\item{space_num}{number of administrative units}

\item{data}{outcome}

\item{space_var}{define the administrative unit variable}

\item{...}{more spatial covariates}

\item{week_var}{define the week time variable. it must be in aweek::date2week(date, week_start = "Sunday", floor_day = TRUE) format (e.g. "2018-W51")}
}
\value{
retorno
}
\description{
Create metadata and filter by time and space
}
\section{Functions}{
\itemize{
\item \code{sitrep_delimit_metadata}: cool to create metadata!

\item \code{sitrep_delimit_spacetime_eval}: create 2 (deprecated) -> now is ..._count

\item \code{sitrep_delimit_spacetime_make}: create 3  (deprecated) -> now is ..._filter

\item \code{sitrep_delimit_spacetime_eval_free}: create 2 free (deprecated) -> now is ..._count_free

\item \code{sitrep_delimit_spacetime_count}: create 2

\item \code{sitrep_delimit_spacetime_filter}: create 3

\item \code{sitrep_delimit_spacetime_count_free}: create 2 free
}}

\examples{

library(outbreaks)
library(tidyverse)
library(aweek)
library(sitreper)
#data(package="outbreaks")

mers_korea_2015 \%>\% summary()

mers_korea_2015_linelist <- mers_korea_2015 \%>\%
  pluck("linelist") \%>\%
  as_tibble() \%>\%

  # formats for sitreper
  mutate(
    epiweek=date2week(dt_onset, week_start = "Sunday", floor_day = TRUE),
    epiweek_fct=as.factor(epiweek),
    loc_hosp_chr=as.character(loc_hosp)
  )

mers_korea_2015_linelist \%>\% nrow()
mers_korea_2015_linelist \%>\% count(epiweek)
mers_korea_2015_linelist \%>\% count(loc_hosp,sort = T)

mers_korea_2015_linelist \%>\%
  sitrep_delimit_spacetime_filter(week_var = epiweek_fct,
                                  week_ini = "2015-W20",
                                  week_fin = "2015-W21",
                                  space_var = loc_hosp_chr,
                                  space_num = 4)

mers_korea_2015_linelist \%>\%
  sitrep_delimit_spacetime_count_free(sex,
                                      age_class,
                                      week_var = epiweek_fct,
                                      week_ini = "2015-W22",
                                      week_fin = "2015-W23",
                                      space_var = loc_hosp_chr,
                                      space_num = Inf)


}
